Feature/client/stat

+++ SOMS/service/container/deployment/deployment.service.go
@@ -41,10 +41,7 @@ func (s *DeploymentService) GetOneDeployment(id string) (*deployment.DeploymentR
 }
 
 func (s *DeploymentService) CreateDeployment(n deployment.DeploymentDto) error {
-	_, err := s.Repository.InsertDeployment(n)
-	if err != nil {
-		return err
-	}
+
 	templateStr := `
 apiVersion: {{.ApiVersion}}
 kind: {{.Kind}}
@@ -71,10 +68,14 @@ spec:
         - containerPort: {{.Spec_template_spec_containers_ports_containerPort}}
 `
 
-	fileName := fmt.Sprintf("../../k8s/test/%s_deployment.yaml", n.Metadata_name)
+	fileName := fmt.Sprintf("k8s/test/%s_deployment.yaml", n.Metadata_name)
 	err1 := createYAMLFile(templateStr, fileName, n)
 	if err1 != nil {
 		return err1
+	}
+	_, err := s.Repository.InsertDeployment(n)
+	if err != nil {
+		return err
 	}
 
 	return applyKubectl(fileName)
@@ -91,9 +92,13 @@ func (s *DeploymentService) DeleteDeployment(id string) error {
 	if err0 != nil {
 		return fmt.Errorf("해당 데이터가 없음: %v", err0)
 	}
-	_, err := s.Repository.DeleteOneDeployment(id)
+
 	cmd := exec.Command("kubectl", "delete", "deployment", dpData.Metadata_name)
 	output, err2 := cmd.CombinedOutput()
+	if err2 != nil {
+		return fmt.Errorf("삭제실패: %v", err2)
+	}
+	_, err := s.Repository.DeleteOneDeployment(id)
 	if err != nil {
 		fmt.Print(output)
 		return fmt.Errorf("kubectl 명령 실행 중 오류 발생: %v", err2)
